move resolve storage location function to analysis.py from utils.py
